--- src/assetloading/mod.rs
@@ -1,1 +1,0 @@
-pub mod asset_loader;
