+++ src/assetmanagement/LICENSE
