+++ src/assetmanagement/asset_loader.rs
