+++ src/assetmanagement/mod.rs
@@ -1,0 +1,4 @@
+mod asset_loader;
+pub mod util;
+
+pub use self::asset_loader::AssetManager;
