+++ src/systems/mod.rs
@@ -1,7 +1,5 @@
 mod camera_movement;
 mod ground_reveal;
-mod tile_update;
 
 pub use self::camera_movement::CameraMovementSystem;
 pub use self::ground_reveal::GroundRevealSystem;
-pub use self::tile_update::*;
