--- src/systems/tile_update.rs
@@ -1,94 +1,0 @@
-use amethyst::assets::{AssetStorage, Loader};
-use amethyst::core::cgmath::{Deg, Vector3};
-use amethyst::core::specs::prelude::{Read, ReadExpect, ReadStorage, System, Write, WriteStorage};
-use amethyst::core::transform::Transform;
-use amethyst::renderer::{Material, MaterialDefaults, Mesh, MeshHandle, Texture};
-use assetloading::asset_loader::AssetManager;
-use entities::Tile;
-use level::LevelGrid;
-
-pub struct TileUpdateSystem;
-
-impl<'a> System<'a> for TileUpdateSystem {
-    type SystemData = (
-        Write<'a, TileUpdateQueue>,
-        //Write<'a, UpdateQueue<'a>>,
-        Read<'a, Vec<([[Tile; 3]; 3], String)>>,
-        Read<'a, LevelGrid>,
-        ReadStorage<'a, Tile>,
-        ReadExpect<'a, Loader>,
-        WriteStorage<'a, Transform>,
-        (
-            Write<'a, AssetManager<Mesh>>,
-            WriteStorage<'a, MeshHandle>,
-            Write<'a, AssetStorage<Mesh>>,
-        ),
-        (
-            Write<'a, AssetManager<Texture>>,
-            WriteStorage<'a, Material>,
-            Write<'a, AssetStorage<Texture>>,
-            ReadExpect<'a, MaterialDefaults>,
-        ),
-    );
-
-    fn run(
-        &mut self,
-        (
-            mut tiles,
-            //mut update_queue,
-            dict,
-            level_grid,
-            tile_storage,
-            amethyst_loader,
-            mut transform_storage,
-            (mut mesh_manager, mut mesh_handles, mut mesh_storage),
-            (mut tex_manager, mut mat_storage, mut tex_storage, mat_defaults),
-        ): Self::SystemData,
-    ) {
-        for (x, y) in tiles.drain(..) {
-            // get sprite definition
-            let (wall_type, wall_rotation) =
-                level_grid.determine_sprite_for(x, y, &dict, &tile_storage);
-
-            let mut transform = Transform::default();
-            transform.set_position(Vector3 {
-                x: x as f32,
-                y: 0.0,
-                z: y as f32,
-            });
-            transform.rotate_local(Vector3::new(0.0, 1.0, 0.0), Deg(-wall_rotation as f32));
-
-            let entity = level_grid.get(x as i32, y as i32).unwrap();
-            transform_storage.insert(entity, transform).unwrap();
-            //update_queue.push((entity, wall_type));
-        }
-    }
-}
-
-///
-/// Wrapper for a Vec of Tiles that needs to be updated
-///
-pub struct TileUpdateQueue {
-    tiles: Vec<(i32, i32)>,
-}
-
-impl std::ops::Deref for TileUpdateQueue {
-    type Target = Vec<(i32, i32)>;
-    fn deref(&self) -> &Self::Target {
-        &self.tiles
-    }
-}
-
-impl std::ops::DerefMut for TileUpdateQueue {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.tiles
-    }
-}
-
-impl Default for TileUpdateQueue {
-    fn default() -> Self {
-        TileUpdateQueue {
-            tiles: Vec::<(i32, i32)>::new(),
-        }
-    }
-}
